찐찐찐막!! ʕ⁎̯͡⁎ʔ༄

+++ main.c
@@ -15,35 +15,53 @@ int main(void){
 		menu = selectMenu();
     		if(menu == 0) break;
     		if(menu == 1){
-			listItem(ilist, index);
-			readItem(ilist[0]);
+			if(count == 0){
+				printf("데이터가 존재하지 않습니다...\n");
+			}
+			else{
+				listItem(ilist, index);
+				readItem(ilist[0]);
+			}
+			
     		}
     		else if(menu == 2){
 			count += createItem(&ilist[index]);
 			index++;
     		}
     		else if(menu == 3){
-			int no = selectDataNum(ilist, index);
-			if(no > 0){
-				updateItem(&ilist[no-1]);
-      			}
+			if(count == 0){
+				printf("데이터가 존재하지 않습니다...\n");
+			}
 			else{
-				printf("=> 취소되었습니다!:->\n");
-        			continue;
-      			}
+				int no = selectDataNum(ilist, index);
+				if(no > 0){
+					updateItem(&ilist[no-1]);
+      				}
+				else{
+					printf("=> 취소되었습니다!:->\n");
+        				continue;
+      				}
+			}
     		}
     		else if(menu == 4){
-			int no = selectDataNum(ilist, index);
-			if(no > 0) {
-				printf("정말로 삭제하시겠습니까?(삭제: 1) ");
-      			}
-			int delok;
-			scanf("%d", &delok);
-			if(delok == 1) {
-				deleteItem(&ilist[no-1]);
-        			count--;
-        			printf("=> 삭제되었습니다 :-)\n");
-      			}
+			if(count == 0){
+				printf("데이터가 존재하지 않습니다...\n");
+			}
+			else{
+				int no = selectDataNum(ilist, index);
+				if(no > 0) {
+					printf("정말로 삭제하시겠습니까?(삭제: 1) ");
+      				}
+				int delok;
+				scanf("%d", &delok);
+				if(delok == 1) {
+					deleteItem(&ilist[no-1]);
+        				count--;
+        				printf("=> 삭제되었습니다 :-)\n");
+      				}
+			}
+				
+			
    		}
     		else if(menu == 5) {
 			saveData(ilist, index);
